Update description to be more verbose
The intent here isn't to make this in to a marketing piece, rather to
provide a succinct, yet also sufficiently descriptive snapshot
overview/introduction to what laminas-hydrator is. It's aimed at people
who are seeing the package/component for the first time and need to
quickly know what it is. I hope that this change does that, and gives
them incentive to delve deeper.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>