Rework the Quick Start guide
The intent of this change is to make the content far more accessible to
first time readers. The changes include:

1. Update the introductory content to build upon the core description in
index.md. This time, the user is a step deeper, and seeking to learn
more. So this description takes them that bit deeper, providing them
quick and easy access to the next level down in to the component.
2. As they're still new to the package, it brings the usage section
right up near the top, so that they can get a sample of how the package
works and what they can expect.
3. It moves the available implementations to after the usage section,
simplifies the headers and descriptions of each one, and adds a usage
example for any that don't have one. The reason for this is that, to me,
using the FQN as the section header is redundant and too verbose. The
descriptions were reworked to read a little more like a narrative, yet
still maintain the focus of a core documentation guide. I tried hard to
make the text flow, yet still be active and to the point. There should
be usage examples for each one, as you can never be sure which one the
user may be most interested in, which one they need, and for
completeness sake. I also sorted the hydrators in to what seemed to be
the most applicable order of complexity, from the most basic to the most
advanced.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>