fix: ensure hydrators work with anonymous objects
This patch modifies the `FilterInterface` to now include a second, optional argument, `?object $instance = null`.

Previously, filters already checked to see if static property notation was provided or not.
This extends that slightly to allow passing an optional second argument, a `null` or `object` value representing the instance being extracted or hydrated.
For those filters that then use reflection, the `$instance` can be used as the primary reflection argument, and the `$property` argument as the second.
This approach allows the `ClassMethodsHydrator` to filter properly based on the specific instance under scrutiny.

Additionally, this makes changes to `ClassMethodsHydrator` to:

- Detect if an anonymous class was used.
- Use the `spl_object_hash()` of that instance for caching the method list.
- Pass the `$object` as the `$instance` argument to filters.

I've added tests to each of the hydrators to verify they work as expected when presented with anonymous objects.

Finally, I've also provided other code updates as flagged by Psalm, based on functionality touched from the changes I made for the patch.

Fixes #25
Fixes #26

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>